requirements.txt
albumentations~=1.1.0
matplotlib~=3.5.2
numpy~=1.22.3
opencv_python_headless~=4.5.5.64
pandas~=1.4.2
Pillow~=9.1.0
pycocotools~=2.0.4
PyYAML~=6.0
setuptools~=59.5.0
torch~=1.10.2
torchvision~=0.11.3
tqdm~=4.63.2
torchsummary~=1.5.1
tensorboard~=2.8.0
